Practica 2 - Arreglos final, solucion a 0/0

+++ LabNetPractica2/LabNetPractica2/Application/ExtensionMethods/IntExtension.cs
@@ -4,7 +4,7 @@
     {
         public static double Divide(this double numerator, double denominator)
         {
-            return (double)numerator / denominator;
+            return numerator / denominator;
         }
     }
 }
